backend/requirements.txt
# =========================================
# PhishGuard — Full Requirements
# Covers: app.py, rag_engine.py,
#         vector_store.py, feature.py
# =========================================
# Install:
#   pip install -r requirements.txt

# ── Web Framework ─────────────────────────
flask
flask-cors
gunicorn

# ── ML & Data Science ─────────────────────
scikit-learn
xgboost
numpy
pandas
scipy
joblib

# ── NLP / Text Processing ─────────────────
beautifulsoup4
lxml

# ── Plotting ──────────────────────────────
matplotlib
seaborn

# ── URL & Network ─────────────────────────
requests
urllib3
python-whois
googlesearch-python
python-dateutil

# ── Gemini / Google AI ────────────────────
google-genai
google-generativeai

# ── RAG / Vector Store ────────────────────
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
huggingface-hub

# ── Environment ───────────────────────────
python-dotenv